+++ upcloud/testdata/upcloud_server_group/step2.tf
@@ -16,8 +16,12 @@ resource "upcloud_server" "test" {
 resource "upcloud_server_group" "tf_test_1" {
   title                = "tf_test_1_updated"
   anti_affinity_policy = "strict"
-  labels               = {
+  labels = {
     "key1" = "val1"
     "key2" = "val2"
   }
+}
+
+resource "upcloud_server_group" "tf_test_2" {
+  title = "tf_test_2"
 }